Feature - enabled backend with flags
This commit adds the possibility of stipulating
which backend should be compiled into the
library.

--with-sqlite3 will link the sqlite3 library.
--with-postgres will link the libpq library.

Both can be added at the same time and the
solution was tested.

Also the libraries are now found via pkg-config.
This has been added to the configure.ac script.

In order to have this working, one should call
the autogen.sh script once which will re-generate
the Makefile.in, aclocal.m4, compile, config.h.in,
configure, depcomp, install-sh and missing files.

Re-generate the autoconf-files:
./autogen.sh --with-sqlite3 --with-postgres --prefix $INSTALLDIR.

After this one can call "make all && make install"